Separate writer from classes
