+++ resim/metrics/python/BUILD
@@ -2,10 +2,20 @@ load("@resim_python_deps//:requirements.bzl", "requirement")
 load("@rules_python//python:defs.bzl", "py_binary", "py_library", "py_test")
 
 py_library(
+    name = "metrics",
+    srcs = ["metrics.py"],
+    deps = [
+        ":metrics_utils",
+        "//resim/metrics/proto:metrics_proto_py",
+    ],
+)
+
+py_library(
     name = "metrics_writer",
     srcs = ["metrics_writer.py"],
     deps = [
         requirement("numpy"),
+        ":metrics",
         ":metrics_utils",
         "//resim/metrics/proto:metrics_proto_py",
         "@com_google_protobuf//:protobuf_python",
@@ -28,6 +38,7 @@ py_binary(
     srcs = ["metrics_example.py"],
     deps = [
         requirement("numpy"),
+        ":metrics",
         ":metrics_utils",
         ":metrics_writer",
         "//resim/metrics/proto:metrics_proto_py",
@@ -41,6 +52,7 @@ py_test(
     srcs = ["metrics_test.py"],
     deps = [
         requirement("numpy"),
+        ":metrics",
         ":metrics_writer",
         "//resim/metrics/proto:metrics_proto_py",
         "//resim/utils/proto:uuid_proto_py",
